+++ utils/storage.py
@@ -28,9 +28,12 @@ def local_storage_path(base_dir: Path, ds: Dataset) -> Path:
     dest_path = Path(
         base_dir,
         str(ds.PatientName),
-        "Study" + "-" + ds.AccessionNumber + "-" + study_uid_suffix,
-        "Series" + "-" + str(ds.SeriesNumber),
-        "Image" + "-" + str(ds.InstanceNumber) + ".dcm",
+        ds.StudyInstanceUID,
+        ds.SeriesInstanceUID,
+        ds.SOPInstanceUID + ".dcm",
+        # "Study" + "-" + ds.AccessionNumber + "-" + study_uid_suffix,
+        # "Series" + "-" + str(ds.SeriesNumber),
+        # "Image" + "-" + str(ds.InstanceNumber) + ".dcm",
     )
     # Ensure all directories in the path exist
     os.makedirs(os.path.dirname(dest_path), exist_ok=True)
@@ -63,30 +66,21 @@ def count_studies_series_images(patient_path: str):
     return study_count, series_count, image_count
 
 
-def images_stored(base_dir: Path, anon_pt_id: str, study_uid: str, acc_no: str) -> int:
+def count_study_images(base_dir: Path, anon_pt_id: str, study_uid: str) -> int:
     """
     Counts the number of images stored in a given study directory.
 
     Args:
         anon_uid (str): The anonymous patient ID.
         study_uid (str): The study UID.
-        acc_no (str): The accession number.
 
     Returns:
         The number of images stored in the study directory.
     """
-    ds: Dataset = Dataset()
-    ds.StudyInstanceUID = study_uid
-    ds.PatientID = anon_pt_id
-    ds.PatientName = anon_pt_id
-    ds.AccessionNumber = acc_no
-    ds.SeriesNumber = 1
-    ds.InstanceNumber = 1
-    instance_path = local_storage_path(base_dir, ds)
-    study_path = instance_path.parent.parent
+    study_path = Path(base_dir, anon_pt_id, study_uid)
     image_count = 0
 
-    for root, dirs, files in os.walk(study_path):
+    for _, _, files in os.walk(study_path):
         for file in files:
             if file.endswith(".dcm"):
                 image_count += 1